rename package to memoization, so that we can patch more easily
