+++ lib/streamlit/__init__.py
@@ -85,7 +85,7 @@ from streamlit.proto import ForwardMsg_pb2 as _ForwardMsg_pb2
 # Modules that the user should have access to. These are imported with "as"
 # syntax pass mypy checking with implicit_reexport disabled.
 from streamlit.caching import cache as cache
-from streamlit.memo.memo import memo as memo
+from streamlit.memoization.memo import memo as memo
 
 # This is set to True inside cli._main_run(), and is False otherwise.
 # If False, we should assume that DeltaGenerator functions are effectively
